Initial version of a very simple yet extensible Gherkin runner
The goal is not to provide a full Cucumber/SpecFlow replacement,
but to allow augmenting a unit test by providing extensible step
execution APIs that can be easily reused to leverage Gherkin
spec files while keeping the unit tests mostly familiar.

See Sample.feature for a basic test and Sample.cs for unit tests
that do not extend the library, and CustomSample.feature and
CustomSample.cs for an example that does extend it with conventional
behavior for `Given key=value` like state setting that can easily
be checked in a `Then` assertion.