Drastically simplify implementation
The separation of builder + runner was only causing friction
for very little gain. Creating derived scenario and context
types was more important and should be simpler.

Instead of myriad extension methods, it's better to just rely
on the inherited types to expose an API that is intuitive to
the type of scenario being run (i.e. BuildScenario).